making arm positions millis instead of meters (#70)

+++ kinematics/nloptInverseKinematics_test.go
@@ -21,9 +21,9 @@ func TestCreateNloptIKSolver(t *testing.T) {
 
 	pos := pb.ArmPosition{X: 170, Y: 0, Z: 180, RX: 0, RY: 0, RZ: 0}
 	transform := kinmath.NewQuatTransFromRotation(pos.RX, pos.RY, pos.RZ)
-	transform.SetX(pos.X)
-	transform.SetY(pos.Y)
-	transform.SetZ(pos.Z)
+	transform.SetX(float64(pos.X))
+	transform.SetY(float64(pos.Y))
+	transform.SetZ(float64(pos.Z))
 
 	ik.AddGoal(transform, 0)
 
@@ -34,9 +34,9 @@ func TestCreateNloptIKSolver(t *testing.T) {
 
 	pos = pb.ArmPosition{X: -46, Y: -133, Z: 372, RX: -18, RY: -33, RZ: -11}
 	transform = kinmath.NewQuatTransFromRotation(pos.RX*math.Pi/180, pos.RY*math.Pi/180, pos.RZ*math.Pi/180)
-	transform.SetX(pos.X / 2)
-	transform.SetY(pos.Y / 2)
-	transform.SetZ(pos.Z / 2)
+	transform.SetX(float64(pos.X / 2))
+	transform.SetY(float64(pos.Y / 2))
+	transform.SetZ(float64(pos.Z / 2))
 
 	ik.AddGoal(transform, 0)
